# Conflicts:
#	src/main/java/com/SerieTemporel/Service/UtilisateurService.java
#	src/main/java/com/SerieTemporel/modele/Utilisateur.java
#	src/main/java/com/SerieTemporel/repository/EvenementRepo.java
#	src/main/java/com/SerieTemporel/repository/SerieRepo.java
#	src/main/java/com/SerieTemporel/repository/UtilisateurRepo.java

+++ src/main/java/com/SerieTemporel/repository/EvenementRepo.java
@@ -7,7 +7,7 @@ import com.SerieTemporel.modele.Evenement;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EvenementRepo extends JpaRepository<Evenement, Integer> {
+public interface EvenementRepo extends JpaRepository<Evenement, Long> {
     
 }
